+++ src/attune/__init__.py
@@ -55,7 +55,7 @@ Copyright 2025 Smart AI Memory, LLC
 Licensed under the Apache License, Version 2.0
 """
 
-__version__ = "3.0.3"
+__version__ = "3.0.4"
 __author__ = "Patrick Roebuck"
 __email__ = "patrick.roebuck@smartaimemory.com"